+++ site/docs/markdown.md
@@ -4,7 +4,7 @@ description: Helpers for converting markdown to HTML
 order: 15
 ---
 
-> **Availability:** Registered in the Node build only. Not available in the browser build.
+> **Availability:** Registered in both the Node and browser builds. The browser build renders inline markdown only — the Node build also accepts a file path and will read its contents before rendering.
 
 ## markdown
