+++ src/helper-registry.ts
@@ -19,7 +19,7 @@ import { helpers as inflectionHelpers } from "./helpers/inflection.js";
 import { helpers as loggingHelpers } from "./helpers/logging.js";
 import { helpers as matchHelpers } from "./helpers/match.js";
 import { helpers as mathHelpers } from "./helpers/math.js";
-import { helpers as mdHelpers } from "./helpers/md.js";
+import { helpers as mdHelpers } from "./helpers/md.node.js";
 import { helpers as miscHelpers } from "./helpers/misc.js";
 import { helpers as numberHelpers } from "./helpers/number.js";
 import { helpers as objectHelpers } from "./helpers/object.js";
